Refactoring k8s deployment files
