--- Daratos_API/flask/app/__init__.py
@@ -1,5 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-from app import api_base
