+++ Daratos_API/flask/flask-deployment.yaml
@@ -1,0 +1,30 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: flask-app
+  labels:
+    name: flask-app
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      name: flask-app
+  template:
+    metadata:
+      name: flask-app
+      labels:
+        name: flask-app
+    spec:
+      containers:
+        - name: flask-app
+          image: gcr.io/daratos/flask-app:v1
+          ports:
+            - containerPort: 8080
+          resources:
+            requests:
+              memory: 256Mi
+            limits:
+              memory: 512Mi
+          env:
+            - name: DEBUG_MODE
+              value: "1"
