+++ Daratos_API/flask/requirements.txt
@@ -3,5 +3,10 @@ nltk==3.4.5
 numpy==1.16.2
 tensorflow==1.14.0
 firebase-admin==3.1.0
-uWSGI==2.0.18
 keras==2.1.5
+Click==7.0
+gunicorn==19.9.0
+itsdangerous==1.1.0
+Jinja2==2.10
+MarkupSafe==1.1.0
+Werkzeug==0.14.1
