--- Daratos_API/flask/run.py
@@ -1,7 +1,0 @@
-from app import app
-import os
-
-app.config['ENV'] = "development"
-
-if __name__ == "__main__":
-    app.run(use_reloader=False)
