+++ Daratos_API/flask/text_processor.py
